# services/invoice-gateway-api/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice Gateway API
  version: 1.0.0
  description: |
    Central entry point for all invoice submissions to the eRačun platform.
    Supports XML, JSON, and PDF invoice formats with async processing.
  contact:
    name: eRačun Team
    email: api@eracun.hr

servers:
  - url: http://localhost:3001
    description: Development
  - url: https://staging.eracun.internal
    description: Staging
  - url: https://api.eracun.hr
    description: Production

tags:
  - name: invoices
    description: Invoice submission and status
  - name: health
    description: Health check endpoints

paths:
  /api/v1/invoices:
    post:
      tags:
        - invoices
      summary: Submit invoice for processing
      description: |
        Submit an invoice for validation and processing. The invoice will be
        queued for async processing through the 6-layer validation pipeline.
      operationId: submitInvoice
      parameters:
        - name: X-Idempotency-Key
          in: header
          required: true
          description: UUID for idempotent request handling
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        - name: X-Request-ID
          in: header
          required: false
          description: Request correlation ID for tracing
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
              format: binary
              description: UBL 2.1 XML invoice
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceSubmission'
          application/pdf:
            schema:
              type: string
              format: binary
              description: PDF invoice (requires OCR processing)
      responses:
        '202':
          description: Invoice accepted for processing
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Request correlation ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceAccepted'
        '400':
          description: Bad request (validation failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Duplicate submission (idempotency key conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceAccepted'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/invoices/{invoiceId}:
    get:
      tags:
        - invoices
      summary: Get invoice status
      description: Retrieve the processing status of a submitted invoice
      operationId: getInvoiceStatus
      parameters:
        - name: invoiceId
          in: path
          required: true
          description: Invoice ID returned from submission
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invoice status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceStatus'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/health:
    get:
      tags:
        - health
      summary: Health check
      description: Check service health and dependency status
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
        '503':
          description: Service is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /api/v1/health/ready:
    get:
      tags:
        - health
      summary: Readiness check
      description: Check if service is ready to accept traffic
      operationId: readinessCheck
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready:
                    type: boolean
                    example: true
        '503':
          description: Service is not ready

  /api/v1/health/live:
    get:
      tags:
        - health
      summary: Liveness check
      description: Check if service is alive (for Kubernetes)
      operationId: livenessCheck
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  alive:
                    type: boolean
                    example: true

components:
  schemas:
    InvoiceSubmission:
      type: object
      required:
        - invoiceNumber
        - issueDate
        - supplier
        - buyer
        - lineItems
        - amounts
      properties:
        invoiceNumber:
          type: string
          example: "INV-2025-001"
        issueDate:
          type: string
          format: date
          example: "2025-11-14"
        dueDate:
          type: string
          format: date
          example: "2025-12-14"
        supplier:
          $ref: '#/components/schemas/Party'
        buyer:
          $ref: '#/components/schemas/Party'
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
        amounts:
          $ref: '#/components/schemas/Amounts'

    Party:
      type: object
      required:
        - name
        - address
        - vatNumber
      properties:
        name:
          type: string
          example: "Test d.o.o."
        address:
          $ref: '#/components/schemas/Address'
        vatNumber:
          type: string
          pattern: '^HR\d{11}$'
          example: "HR12345678901"
        email:
          type: string
          format: email
        phone:
          type: string

    Address:
      type: object
      required:
        - street
        - city
        - postalCode
        - country
      properties:
        street:
          type: string
          example: "Ilica 1"
        city:
          type: string
          example: "Zagreb"
        postalCode:
          type: string
          example: "10000"
        country:
          type: string
          pattern: '^[A-Z]{2}$'
          example: "HR"

    LineItem:
      type: object
      required:
        - id
        - description
        - quantity
        - unitPrice
        - kpdCode
        - vatRate
      properties:
        id:
          type: string
          example: "line-1"
        description:
          type: string
          example: "Professional services"
        quantity:
          type: number
          minimum: 0
          example: 1
        unit:
          type: string
          example: "EA"
        unitPrice:
          type: number
          minimum: 0
          example: 1000.00
        kpdCode:
          type: string
          pattern: '^\d{6}$'
          example: "123456"
        vatRate:
          type: number
          enum: [0, 5, 13, 25]
          example: 25

    Amounts:
      type: object
      required:
        - net
        - vat
        - gross
        - currency
      properties:
        net:
          type: number
          minimum: 0
          example: 1000.00
        vat:
          type: array
          items:
            $ref: '#/components/schemas/VATBreakdown'
        gross:
          type: number
          minimum: 0
          example: 1250.00
        currency:
          type: string
          enum: [EUR, HRK]
          example: "EUR"

    VATBreakdown:
      type: object
      properties:
        rate:
          type: number
          enum: [0, 5, 13, 25]
        base:
          type: number
        amount:
          type: number
        category:
          type: string
          enum: [STANDARD, REDUCED, SUPER_REDUCED, EXEMPT]

    InvoiceAccepted:
      type: object
      properties:
        invoiceId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        status:
          type: string
          enum: [QUEUED, PROCESSING]
          example: "QUEUED"
        trackingUrl:
          type: string
          format: uri
          example: "https://api.eracun.hr/api/v1/invoices/123e4567-e89b-12d3-a456-426614174000"
        acceptedAt:
          type: string
          format: date-time

    InvoiceStatus:
      type: object
      properties:
        invoiceId:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [QUEUED, PROCESSING, VALIDATING, VALIDATED, TRANSFORMING, SUBMITTING, COMPLETED, FAILED]
        progress:
          type: object
          properties:
            currentStep:
              type: string
            totalSteps:
              type: integer
            percentage:
              type: integer
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        validationResult:
          type: object
        error:
          $ref: '#/components/schemas/Error'

    HealthStatus:
      type: object
      properties:
        status:
          type: string
          enum: [UP, DOWN, DEGRADED]
          example: "UP"
        timestamp:
          type: string
          format: date-time
        dependencies:
          type: object
          additionalProperties:
            type: string
            enum: [UP, DOWN]
          example:
            database: "UP"
            rabbitmq: "UP"
            redis: "UP"
        version:
          type: string
          example: "1.0.0"

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "ERR_1001"
            message:
              type: string
              example: "Invalid OIB format"
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            correlationId:
              type: string
              format: uuid